Some documentation / refactoring as OpenSRF API learning occurs
